Fix xcode and deprecation warnings
