+++ Provenance/Utilities/PVLogViewController.xib
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.iPad.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch.iPad" propertyAccessControl="none" useAutolayout="YES" colorMatched="YES">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.iPad.XIB" version="3.0" toolsVersion="19158" targetRuntime="iOS.CocoaTouch.iPad" propertyAccessControl="none" useAutolayout="YES" colorMatched="YES">
     <device id="ipad9_7" orientation="portrait" layout="fullscreen" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19141"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
@@ -13,7 +13,6 @@
                 <outlet property="contentView" destination="ir3-aG-kbV" id="1s7-NQ-ZtQ"/>
                 <outlet property="doneButton" destination="20" id="34"/>
                 <outlet property="logListButton" destination="BF8-Bz-b6R" id="EPj-zY-9Qr"/>
-                <outlet property="searchDisplayController" destination="22" id="24"/>
                 <outlet property="segmentedControl" destination="17" id="31"/>
                 <outlet property="textView" destination="4" id="14"/>
                 <outlet property="toolbar" destination="7" id="12"/>
@@ -26,7 +25,7 @@
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
                 <toolbar opaque="NO" clearsContextBeforeDrawing="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="7">
-                    <rect key="frame" x="0.0" y="0.0" width="768" height="44"/>
+                    <rect key="frame" x="0.0" y="20" width="768" height="44"/>
                     <constraints>
                         <constraint firstAttribute="height" constant="44" id="REJ-bi-77z"/>
                     </constraints>
@@ -34,11 +33,10 @@
                         <barButtonItem style="plain" systemItem="flexibleSpace" id="16"/>
                         <barButtonItem style="plain" id="18">
                             <segmentedControl key="customView" opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="top" segmentControlStyle="bar" selectedSegmentIndex="0" id="17">
-                                <rect key="frame" x="246.5" y="6" width="275" height="32"/>
+                                <rect key="frame" x="246.5" y="5.5" width="275" height="33"/>
                                 <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                                 <segments>
                                     <segment title="Live Log"/>
-                                    <segment title="System Log"/>
                                     <segment title="File Logs"/>
                                 </segments>
                                 <connections>
@@ -59,11 +57,11 @@
                     </connections>
                 </toolbar>
                 <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="ir3-aG-kbV" userLabel="contentview">
-                    <rect key="frame" x="0.0" y="44" width="768" height="980"/>
+                    <rect key="frame" x="0.0" y="64" width="768" height="960"/>
                     <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                 </view>
                 <textView clipsSubviews="YES" contentMode="scaleToFill" showsHorizontalScrollIndicator="NO" editable="NO" translatesAutoresizingMaskIntoConstraints="NO" id="4">
-                    <rect key="frame" x="0.0" y="44" width="768" height="980"/>
+                    <rect key="frame" x="0.0" y="64" width="768" height="960"/>
                     <string key="text">Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.</string>
                     <color key="textColor" systemColor="darkTextColor"/>
                     <fontDescription key="fontDescription" type="system" pointSize="12"/>
@@ -92,14 +90,6 @@
             <simulatedStatusBarMetrics key="simulatedStatusBarMetrics" statusBarStyle="blackOpaque"/>
             <point key="canvasLocation" x="118" y="42"/>
         </view>
-        <searchDisplayController id="22">
-            <connections>
-                <outlet property="delegate" destination="-1" id="28"/>
-                <outlet property="searchContentsController" destination="-1" id="25"/>
-                <outlet property="searchResultsDataSource" destination="-1" id="26"/>
-                <outlet property="searchResultsDelegate" destination="-1" id="27"/>
-            </connections>
-        </searchDisplayController>
         <barButtonItem title="Select File" id="BF8-Bz-b6R">
             <connections>
                 <action selector="logListButtonClicked:" destination="-1" id="OLZ-mk-o1e"/>
