+++ ProvenanceTV/PVSearchViewController.swift
@@ -94,7 +94,7 @@ final class PVSearchViewController: UICollectionViewController, GameLaunchingVie
             })
             .flatMapLatest({ game in
                 self.gameLibrary.toggleFavorite(for: game)
-                    .catchError { _ in
+                    .catch { _ in
                         ELOG("Failed to toggle Favourite for game \(game.title)")
                         return .never()
                 }
